DduCore and DVD Cover interface

+++ SRC/Hardware/AI.cpp
@@ -342,7 +342,7 @@ static void __fastcall write_in_mbox_l(uint32_t addr, uint32_t data) { DBHalt("P
 void AIUpdate()
 {
     // *** update audio DMA ***
-    if((Gekko::Gekko->GetTicks() >= ai.dmaTime) && (ai.len & AID_EN))
+    if(((uint64_t)Gekko::Gekko->GetTicks() >= ai.dmaTime) && (ai.len & AID_EN))
     {
         if (ai.dcnt == 0)
         {
